fix(balance): 移除全局移速倍率，怪物移速直接设为旧版×0.75值；设置界面仅显示刷怪倍率；修复auto_destroy不生效BUG

+++ entities/monster.py
@@ -22,7 +22,7 @@ class Monster:
     x: float = 0.0
     y: float = 0.0
     cooldown_remaining: float = 0.0   # 当前剩余冷却（秒）
-    speed: int = MONSTER_SPEED
+    speed: float = MONSTER_SPEED
     alive: bool = True
     aggro: bool = False       # 是否已进入战斗状态
